requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
audeer==2.2.1
audiofile==1.5.1
audmath==1.4.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gradio_client==1.5.4
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-core==0.3.29
langchain-groq==0.2.3
langgraph==0.2.62
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.51
langsmith==0.2.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==1.0.3
msgpack==1.1.0
narwhals==1.21.1
numpy==2.2.1
opencv-python==4.10.0.84
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proglog==0.1.10
protobuf==5.29.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.0
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
websockets==14.1
